requirements.txt
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
decorator==5.2.1
filelock==3.18.0
fsspec==2025.3.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
Jinja2==3.1.6
llvmlite==0.44.0
MarkupSafe==3.0.2
more-itertools==10.6.0
moviepy==2.1.2
mpmath==1.3.0
networkx==3.4.2
numba==0.61.0
numpy==2.1.3
openai-whisper==20240930
opencv-python==4.11.0.86
packaging==24.2
pillow==10.4.0
proglog==0.1.10
pycparser==2.22
pytesseract==0.3.13
python-dotenv==1.1.0
regex==2024.11.6
requests==2.32.3
setuptools==78.1.0
soundfile==0.13.1
srt==3.5.3
sympy==1.13.1
tiktoken==0.9.0
torch==2.6.0
tqdm==4.67.1
typing_extensions==4.13.0
urllib3==2.3.0
vosk==0.3.44
websockets==15.0.1
